got it partially working
